refactor: Refactor from Review CompanyDataStorage logic and remove redundancy
Consolidate the handling of company data storage into CompanyDataStorageHandler methods for improved readability and maintenance. Remove redundant code and streamline the process of populating and setting company data.
