+++ apps/web/src/components/questionnaire/result/list/ResultListInTabs.vue
@@ -20,7 +20,6 @@
       <ResultProjectList
         :filtered-projects="filteredProjects"
         :has-error="hasError"
-        :has-registered-data="hasRegisteredData"
       />
     </DsfrTabContent>
