+++ apps/web/src/stores/usedTrack.ts
@@ -1,11 +1,9 @@
 import { useNavigationStore } from '@/stores/navigation'
 import { useTrackStore } from '@/stores/track'
 import {
-  EstablishmentFront,
   LegalCategory,
   type NextTrackRuleSet,
   type QuestionnaireData,
-  SiretValue,
   StructureSize,
   type Track,
   TrackComponent,
@@ -22,7 +20,6 @@ import Translation from '@/utils/translation'
 import { acceptHMRUpdate, defineStore } from 'pinia'
 import { computed, ref, toRaw } from 'vue'
 import CompanyDataStorage from '@/utils/storage/companyDataStorage'
-import { CompanyDataStorageKey } from '@/types/companyDataType'
 import { CompanyDataStorageHandler } from '@/utils/storage/companyDataStorageHandler'
 
 export const useUsedTrackStore = defineStore('usedTrack', () => {
@@ -69,15 +66,8 @@ export const useUsedTrackStore = defineStore('usedTrack', () => {
       .filter((questionnaireDatum) => questionnaireDatum?.length)
       .flat(1)
 
-    if (_canUseCompanyData(data, CompanyDataStorageKey.Siret)) {
-      data.push(CompanyDataStorage.getSiret() as QuestionnaireData)
-    }
+    CompanyDataStorageHandler.populateCompletedQuestionnaire(data)
 
-    if (_canUseCompanyData(data, CompanyDataStorageKey.Size)) {
-      data.push({ [CompanyDataStorageKey.Size]: CompanyDataStorage.getSize() } as QuestionnaireData)
-    }
-
-    console.log(data)
     return data
   })
 
@@ -124,13 +114,7 @@ export const useUsedTrackStore = defineStore('usedTrack', () => {
       if (value) {
         useNavigationStore().updateSearchParam({ name: current.value.id, value: value })
 
-        if (current.value.id === TrackId.Siret && value !== SiretValue.Wildcard) {
-          CompanyDataStorage.setSiret(selectedOptions[0].questionnaireData as EstablishmentFront)
-        }
-
-        if (current.value.id === TrackId.StructureWorkforce) {
-          CompanyDataStorage.setSize(value as StructureSize)
-        }
+        CompanyDataStorageHandler.setDataFromTrack(current.value.id, value, selectedOptions)
       }
     }
 
@@ -287,22 +271,9 @@ export const useUsedTrackStore = defineStore('usedTrack', () => {
       questionnaireData.onlyEligible = true
     }
 
-    CompanyDataStorageHandler.populateFromStorage(questionnaireData)
+    CompanyDataStorageHandler.populateQuestionnaireData(questionnaireData)
 
     return questionnaireData
-  }
-
-  function _canUseCompanyData(data: (QuestionnaireData | undefined)[], key: CompanyDataStorageKey): boolean {
-    if (CompanyDataStorage.hasItem(key)) {
-      const storageItem = CompanyDataStorage.getItem(key)
-
-      return (
-        CompanyDataStorage.hasItem(key) &&
-        !data.some((item) => item?.[key] === (typeof storageItem === 'string' ? storageItem : (storageItem as QuestionnaireData)?.[key]))
-      )
-    }
-
-    return false
   }
 
   async function setFromNavigation() {
@@ -317,13 +288,7 @@ export const useUsedTrackStore = defineStore('usedTrack', () => {
       const value = useNavigationStore().query[trackId] as string | string[]
       const selectedOptions: TrackOptionsUnion[] = await useTrackStore().getSelectedOptionsByTrackAndValue(track, value)
 
-      if (trackId === TrackId.Siret && selectedOptions.length > 0) {
-        CompanyDataStorage.setSiret(selectedOptions[0].questionnaireData as EstablishmentFront)
-      }
-
-      if (trackId === TrackId.StructureWorkforce) {
-        CompanyDataStorage.setSize(value as StructureSize)
-      }
+      CompanyDataStorageHandler.setDataFromTrack(trackId as TrackId, value, selectedOptions)
 
       if (selectedOptions.length === 0) {
         useNavigationStore().deleteSearchParam(trackId)
